new stuff
I fleshed out the init of a few things. I tried to take a stab at the issue of not printing player controller input. I think what is happening is that the instantiation of the player controller in the camera view creates objects of the "controller" type and not the "PlayerController" type and the actual input processing functions aren't accessed. Will look into this more later- but I think this is the problem