new changes to map creation algorithm
now, instead of stages that are explicitly different sizes, each stage is actually the same size, but each stage (during generation) has a random chance to put some holes in the ice along the edges. As you get further from the center, the area where holes can be generated gets bigger. So this gives the illusion of the stages getting smaller in a more subtle way without the super abrupt changes in size