calculating paths
I have now updated the code so that it only changes states and calls the path calculation function only when it needs to rather then every update tick. There is also code that detects if there is a parent to the walrus' current position (meaning that a path has been calculated from that position) and prints out a message. This is ideally where we would iterate from that parent and move the walrus in that direction.  I will see if I can tackle the movement part of this but if someone has any ideas feel free to go for it